Built-in sockets: _outputs and _wait (#36)
## Built-in sockets
All built-in sockets start with `_`.
### `_outputs` socket
This will allow the user to pass the entire `outputs` of an AiiDA node to the input of another node.

### `_wait` socket
Every node has built-in `_wait` input and output socket. The user will allow the user to see the dependency and running order of nodes in the GUI.

## Support deleting nodes
Clean the hanging link before saving the worktree.