build node on fly for AiiDA component (#50)
This PR allows users to use the AiiDA component (calcfunction, CalcJob, WorkChain, etc.) directly inside the WorkGraph instead of manually building the node every time.

The node's inputs and outputs are automatically generated based on the input and output ports of the AiiDA component.

In the case of `calcfunction`, the default output is `result`. If there is more than one output node, one needs to define the node with the outputs explicitly; thus, it can not be used directly in `nodes.new()`.

`calcfunction` also supports dynamic outputs (that return a dict with many nodes). This is treated as one output (`result`) by default.