Fix socket data type (#53)
This PR tries to support the data type validation. **This is still an experimental feature**.

If the data `identifer` is an AiiDA data class, try to build the socket/property on the fly. Use AiiDA's serialize/deserializer.

- The output of the `Normal` node (not an AiiDA component), is saved into the `base.extras`. The result is serialized and saved to the `base.extras` with the key `nodes__results__node_name`.
- When the work graph is finished, deserialize the data and update the socket's value.

When output the result of a node as the output of the node group, we use ['node1.result', 'result']. However, the data type is set to `General` in this case. This will raise an error if it's an AiiDA node, because, a `General` socket will serialize the data using pickle.

In principle, all data kept in the ctx should be serialized. Then deserialize it when using.