Add builtin PythonJob to run any Python function on a remote computer. (#83)
This PR provides:
- `PythonJob` to run Python function on a remote computer
- `GeneralData` to pickle any Python data as a binary and store it in the database.
- Add `PythonJob` as a built-in node

List the file in the remote folder:
```console
$ ls
aiida.out        inputs.pickle   _scheduler-stderr.txt  script.py
_aiidasubmit.sh  results.pickle  _scheduler-stdout.txt
```

Each node creates a `script.py` file on the remote computer, which includes:
- The function definition.
- Loading inputs from `inputs.pickle`.
- Running the function with the loaded inputs.
- Saving the results into `results.pickle`.

For a `CalcJob`, the input data needs to be an AiiDA data node; however, we don't require the user to install AiiDA. This means we should pass normal Python data as arguments when running the Python function on the remote computer. The `WorkGraphEngine` will handle this data transformation when preparing and launching the `CalcJob`.

All AiiDA data that will be passed to the function should have a `value` attribute, which corresponds to its raw Python data. The `GeneralData`, `Int`, `Float`, `Str`, and `Bool` fulfill this requirement, while `List`, `Dict` and `StructureData` are not.

**Python Version** : since pickle is used to store and load data, the Python version on the remote computer should match the one used in the localhost.

```python
metadata = {
    "options": {
        'custom_scheduler_commands' : 'module load anaconda\nconda activate py3.11\n',
    }
}
```

Inputs for each node are pickled into the `inputs.pickle` file.
Outputs from each node are pickled into the `results.pickle` file.

The parent_folder parameter allows a node to access the output files of a parent node. This feature is particularly useful when you want to reuse data generated by a previous computation in subsequent computations.