use `run_remotely` to set `PythonJob`
When adding a new node, one can set `run_remotely=True` to ask the engine to run the function on a remote computer.
In the inputs, user can set `"computer": "merlin6"` to tell which remote computer to run this job, the engine will try to load `python3@{computer}` code, it will create a new one if not exit.

One can also set the `code_label`, `python_path`, `prepend_text` to specific the code.

In the `metadata.custom_scheduler_commands`, one can add commands to load module and activate conda environment add so on.