Widget adds `to_html` method (#94)
For the moment, it's not possible to render the widget on the docs page directly. This PR adds a to_html method so that we can export the workgraph to an HTML file and then load the HTML file into jupyter-notebook; in this case, the GUI will rendered on the docs page.

We convert the react code to vanilla js so that we can merge all codes into one HTML file.