Support passing output socket to the `nodes` input of `ShellJob` (#105)
In `aiida-shell`, users pass AiiDA `Data` to `ShellJob` using the `nodes` input. The `nodes` port is a `namespace,` thus, it accepts a dictionary input with key and `Data`. However, we can not pass the output of a WorkGraph node, because the job has not run yet, and the output is an output socket instead of an AiiDA `Data`.

This PR allows the user to pass the output socket directly to the `nodes` port. For example, if we pass:
```python
nodes={'expression': job1.outputs["stdout"]}
```
It will:
- create an input socket (`node.expression`) for the node.
- make a link between the output socket (`job.outputs["stdout"]`) and the newly created input socket (`node.expression`).
- WorkGraph engine will handle the inputs when running.