Set the nodes during/after the creation of the task. (#113)
In the `ShellJob`, the nodes port is a namespace, thus dynamic. When the user specifies the `nodes` as a dict, whether the value of the item is AiiDA data, `None`, or an output socket from another node, we create an input socket for it. Thus, users can pass data/links to this socket later.