Add pause/play task (#111)
## Pause a task inside WorkGraph
WorkGraph allows users to control the task more precisely:
- Users know all the task names in the WorkGraph, so they can pause any task before it is submitted.
- Users pause the process when it is instantiated so that they can pause it for sure.

For the moment, since WorkGraph uses `run_get_node` to launch `calcfunction` and `workfunction`, so we can not pause them.

Besides, we only support a running workgraph, so that it can receive the message to pasue/play the tasks. In order to do this, we override the `message_receive` method to receive a custom message.

## Technical details
In AiiDA, a process runs the calculation in the `step` method. Inside the `step` method, it will check whether the process is paused. When instantiating the process, we set it to be paused, so that it will wait the pause state. This pause state can be deleted by the `play` action.