Add a basic error handler to WorkGraph (#124)
To “register” an error handler for a WorkGraph, the user simply defines a function that takes self as its single argument and sets it as the error handler of the WorkGraph. If a task fails, WorkGraph will call the error handlers. If the error handler detects the right error code, it resets the task and “fixes” the inputs. The WorkGraph will run the task with the new inputs.