support namespace output in PythonJob (#129)
## Namespace Output

The `PythonJob` allows users to define namespace outputs. A namespace output is a dictionary with keys and values returned by a function. Each value in this dictionary will be serialized to AiiDA data, and the key-value pair will be stored in the database.

### Why Use Namespace Outputs?

- **Dynamic and Flexible**: The keys and values in the namespace output are not fixed and can change based on the task's execution.
- **Querying**: The data in the namespace output is stored as an AiiDA data node, allowing for easy querying and retrieval.
- **Data Provenance**: When the data is used as input for subsequent tasks, the origin of data is tracked.
### Defining Namespace Outputs

To declare a namespace output, set the `identifier` to `Namespace` in the `outputs` parameter of the `@task` decorator. For example:

```python
@task(outputs=[{"name": "structures", "identifier": "Namespace"}])
def generate_surface_slabs():
    # Function logic to generate surface slabs
    return {"slab1": slab_data1, "slab2": slab_data2}
```