Add TaskCollection for waiting_on and children. (#223)
Implement a `TaskCollection` to store a collection of tasks. This is used for the `waiting_on` and `children` attributes of the Task class. It has `add`, `remove` and `clear` methods. The `add` method` will `append` the task to the collection. When adding or removing a task from the collection, it will check whether the task exists.