Adding a prompt to confirm the deletion of a node (#193)
For the prompt react-bootstrap was used and introduced as import.

The bootrap styles overwrite the `th` css style for the table, so we switched
to styling the header `th` instead of `td` which is also recommended usage for
the header of a table.

This required a change in the backend. Now a dry run for the deletion of a node
can be requested to retrieve a list of nodes that depend on this node that will
be deleted as well.

There was a bug in handling of the delete button click in the data node. After
the deletion the workgraph nodes were fetched instead of the data nodes. This
bug hid that the frontend tests were not properly testing the deletion of a
node.  Deleting the data node only worked because switching the workgraph table
returned an empty table that made the test pass.  The problem with existing
tests was that using the first row locator for a table with more than one
element fails after the deletion as the first row is still present. Now we
check the exact number of row elements and the last row is used for deletion.