Add `monitor` and `awaitable` task decorators to create tasks that to poll some state (#235)
This PR introduces `monitor` and `awaitable` task decorators, which can be particularly useful for tasks that poll some state (e.g.,  the presence of a file or the state of another WorkGraph) at given intervals until a success criterion is met. While polling the state, the task will sleep for a given interval (default 1.0 second, can be changed by user), and relinquish control to the WorkGraph engine, so that it can run other tasks.

I created some builtin Tasks:
- TimeMonitor: start a task until a given time.
- FileMonitor: start a task until a file exits.
- TaskMonitor: start a task until another task from a different workgraph is finished.

The monitor and the Calcjob monitor are different in terms of the scope of monitoring. The Monitor in AiiDA currently works on the Calcjob. It is called within the CalcJob's process, so it can modify its output. This is intra-process monitoring. The Monitor in WorkGraph is a task. The monitor task can monitor any events (e.g. time, file changes, etc.) as long as there is a API. Of course, it can also monitor the AiiDA processes using AiiDA API. It can fetch the state of other AiiDA processes, and it can kill/pause/play the process, but it can not modify the data (outputs) of the process. This is inter-process monitoring.

The `awaitable` decorator allows the user to take complete control of the `asyncio` function.