Save task state info in the ctx for checkpoint (#245)
This PR saves the task state information in the context so that it can be restored from a checkpoint. The state info is also stored separately in the extras of the WorkGraph process node for each task so that we can query it. For the process node, we serialize it so that it can also be saved into the extras. 

In order to interact with the running WorkGraph Engine, we need to use `rpc_send` to send a custom message to the engine.