Pickle the input value if it is a function (#262)
If a Python function is used as a task input, we first serialize it using pickle, then use AiiDA's serializer. We created a `PickledLocalFunction` class to serialize the function, and this class is the same as the `PickledFunction`. We use a different name just to label it, so that later we know we need to deserialize it and get the Python function back, before we pass it as a task input.