Pass list of names for task input/output directly (#343)
Previously, one would always have to pass a list of dictionaries to `Task` `inputs` and `outputs`, at least specifying the `name` key in the dictionary. Now, it's possible to also pass a list of strings, which will internally automatically be converted to a list of dictionaries with the `name` key, making the syntax less verbose.