Migrate workgraph data from `base.extras` to `base.attributes` (#410)
In ``aiida-workgraph``, the ``WorkGraphEngine`` class is responsible for executing workflows. Unlike ``WorkChain``, which requires predefined inputs, outputs, and an execution outline for each workflow, ``WorkGraphEngine`` does not have built-in workflow definitions. Instead, it relies on a dynamic input namespace, ``workgraph_data``, allowing users to provide workflow information at runtime. This approach enables greater flexibility in constructing and executing workflows. In the old design, the `workgraph data` is stored in the `base.extras` of the node. However, there are several drawbacks:

- If the `base.extas` is deleted, one can not restore the WorkGraph
- could have a security issue if the data is modified by other people.

In this PR, we save the `workgraph data` as an attribute of the node (`WorkGraphNode`), when creating the process (`WorkGraphEngine`). This attribute can be modified before the `WorkGraph` finishes.

I also add other attributes that remain updatable during execution to reflect the workflow's progress.

- **State**: Represents the current status of a task.
- **Action**: Specifies interactions such as ``pause`` or ``play``.
- **Process**: Stores the primary key (``pk``) of the process node, as a task may launch multiple processes.

To allow these updates, the ``_updatable_attributes`` list defines the fields that can be modified during execution. This list is managed by the ``Sealable`` class, ensuring that only specified attributes can be changed.