WIP implementing tracked context variables
This PR is working on issue issue #408.

For graphical visualization in the GUI we need some class that tracks
the connection that the context variable is used for. The context itself
is completely agnostic of its access by design. Also the context is an
object of plumpy and is completely unaware of the GUI, therefore in
workgraph we use the `node_graph.socket.NodeSocket` for this purpose.
These are heavily coupled with `node_graph.link.NodeLink` that tracks it
connection to other nodes.

For this implementation we need to create a new type of Sockets with extended
capabilities:

1. We need to track usage and create new sockets internally while the
   user is not necessary aware of it.
```python

def increment(x): return x+1

wg.context.x = aiida.orm.Int(5) # create socket x0 with initial value
t = wg.add_task(increment, x = wg.context.x) # link socket x0 to t.inputs.x
wg.context.x = t.outputs.result # create socket x1 and link to results
t2 = wg.add_task(increment, x = wg.context.x) # link socket x1 to t2.inputs.x
```

2. We need to allow bidirectional usage as input and output
```python
wg.add_link(wg.context.x, t.inputs.x)
…
wg.add_link(t.outputs.x, wg.context.x)
```

Since a socket by itself is not containing any logic how it is
populated, we need also a new `Task` class specifically for managing the
`Socket`. We need to at least implement a wrapper around the sockets to
implement custom `__getattr__`, and `__setattr__` to create Sockets on
access. At first glance, a specialized `NodeSocketNamespace` looks like
the the right way to implemnt this, but on the other hand we want to
implement a very specfic access pattern that that easily can interfere
with the other patterns of `NodeSocketNamespace`. So it seems better to
implement a separated `AccessInterface` in `WorkGraph` or the
specialized `Task` that exposes this `NodeSocketNamespace` to the user.

This requires to touch part multiple parts of the code. As we need to
implement has this new `Task` and `Socket` is handled during execution
(population of socket, and passing of this new socket namespace). The
population of the socket is outline in this WIP but the new socket
namespace is not passed to the execution time.

Furthermore, how this is implemented at the GUI is still an open
question. We have a design ideas, see images in the PR on GitHub.