`WorkGraph` task with proper sub-tasks (#411)
A new `WorkGraphTask` is added to better represent a `WorkGraph` that is added as a task to another `WorkGraph`.

This `WorkGraphTask` has the actual `WorkGraph`, as well as its tasks and links attached as properties, making access of the sub-WG elements simple, and accessible via doc notation, such as `parent_wg.tasks.sub_wg.tasks` and `parent_wg.tasks.sub_wg.tasks.inputs`.

The `deepcopy` operation is performed, as when the `graph_data`  is passed to the `from_dict` `classmethod` of WG, it's modified by `pop`ing the `tasks` and assigning them to `nodes`, which would lead to the `tasks` eventually missing without the `deepcopy`.