Save pickled data as an `updatable` attribute (#412)
WorkGraph allows the user to create a task from the local callables (function, WorkChain). Internally, WorkGraph pickles the local callables as binary. Previously, the binary data is stored as an AiiDA data node (PickledFunciton), which is then used as the WorkGraph input. However, the binary data could be very large, and we should provide an option to choose whether to store it or not. However, since we use it as the input, we can not delete it anymore. This is bad.

In this PR, we are not storing it as an additional AiiDA data node, but as the attributes of the WorkGraph process node. Especially, the attribute is an `updatable` attribute, which we can delete when the WorkGraph finishes (but before sealed).

## NodeExecutor
Additionally, refactor the `NodeExecutor` to keep all necessary information of the executor to be pickled, including the pickled binary data and source code of the local callable. This includes:

- task_executors
- task_error_handlers
- workgraph_error_handlers

`Error handler` is the function that runs when a node fails. It is used in the restart workflow and can recover from a failure.