Implements context namespace with accessor for syntactic sugar
Implemtation a namespace for context variables `ContextSocketNamespace`. Does
not contain much logic for the moment. But in future work more logic is needed
here to manage its handling in the engine.

The `ContextSocketNamespaceAccessor` is an access wrapper around the
`ContextSocketNamespace` for syntactically nicer usage hiding complexity of
the socket namespace from the user.

In principle for the current approach the creation sockets and namespaces is not
needed, since it just converted into to the curly brackets dicts when passed to
the engine. At some point we however need these custom objects to represent them
in the GUI. To reduce the complexity of the approach, this PR does only bring
the syntax wrapper around the context namespace.