Distinguish the sockets of PythonJob built-in from the function (#435)
Since we allow users to add dynamic inputs and outputs for the PythonJob, we need to distinguish the sockets belonging to the function from those belonging to the PythonJob.

In this PR, when adding the sockets for the PythonJob, we set the metadata to indicate that the socket is a built-in PythonJob socket.