Use `pyfunction` for default task (#454)
Previously, no provenance was stored for a `normal` task decorated using `@task()`. If the user wants to store the provenance, they need to switch to `calcfunction` or `PythonJob`. But the downsides are
- `calcfunction` needs AiiDA nodes as input and output
- `PythonJob` runs on a remote computer by uploading/retrieving files, which is overkill for a function that only needed to run locally.

Since `aiida-pythonjob v0.2.2`, a new `pyfunction` is added, which runs the function locally, as well as serializes the inputs/outputs like the PythonJob.

In this PR, we use this `pyfunction` for the normal `task`. So internally, it will transfer the function into a `pyfunction`.


Also fix a bug related with `ctx.