Serialize the group_inputs and ctx (#503)
Serialize group inputs and the context variables when submitting the WorkGraph. To maintain consistency, we use the same serializers as those used by `pyfunction` and `PythonJob`. I also updated the docs. In the documentation, I mentioned a **Data Serialization** section, which does not yet exist. But it will be added soon.