Update map (#510)
This PR adds a new `MapItem` task inside the map zone instead of using the context variable to pass the item. This makes the GUI clear.

In the engine, we also copy this `map_item` task, and set the result of using the item. This also makes the logic in the engine clear.