Refactor `@task.graph_builder ` to `@task.graph` (#554)
Because the `graph_builder` is difficult to use in the context manager approach, for example: How to call the `graph_buider` as a task, i.e., as a black box. This is because the `graph_builder` returns a WorkGraph, while in the context manager approach, one calls every component as a task and expects it to return the outputs. Even though `graph_builder` is successfully used in the node-graph programming approach, it is created before the context manager approach. Thus, in this PR, we refactor it to make it work more consistently inside the context manager approach, while still usable in the node-graph programming approach.

The main changes are:
- the `graph task` returns the WorkGraph outputs, instead of the WorkGraph itself
- provide a `get_graph` method