Fix pickle error of AiiDA process in decorator wrapper function (#568)
Cloudpickle tries to serialize the decorator wrapper function (`_make_wrapper`), its closure and attributes. The AiiDA process (CalcJob), which inherits the `ABCMeta`  is inside the wrapper function, however, they are non-picklable. Actually, only when the callable is a function do we need to pass it to the wrapper function to inspect the function signature and prepare the inputs correctly. If the callable is a CalcJob or WorkChain, we don't need to pass it to the wrapper function. 
This PR solves the issue by checking if the callable is a function before passing it to the wrapper function. If it is not a function, we simply pass None to the wrapper function.