Allow graph-level inputs directly (#580)
Now `wg.run(inputs={"x": 1, "y": 2})` instead of `wg.run(inputs={"graph_inputs": {"x": 1, "y": 2}})`. In the back, it first checks if the provided keys are defined inputs, in which case they are assigned. If not, it checks if they are defined tasks, in which case the inputs are passed to the task.