Use `TaggedValue` to track the graph inputs (#627)
We use the `TaggedValue` class, which transparently behaves like the wrapped value, as well as attach a `_socket` attribute referencing its originating graph input socket. When a `TaggedValue` is passed to a task's socket, a link will be made instead of assigning the value. Since this happens at runtime, it captures complex control flow. This should be robust and cover all possible cases.