Bump aiida-pythonjob to 0.3.5 (#651)
Note, aiida-pythonjob uses the `Dict`, `List` and `ArrayData` from aiida-core directly, instead of its own implementation that wraps them with a `value` attribute (because the wrapped class led to a query issue. So one has to use `get_dict`, `get_list`, but the good thing is that from `aiida-core 2.7.0`, `Dict` and `List` have a `value` attribute. So this change is needed only for the old aiida-core version.

- aiida-pythonjob to 0.3.5
- node-graph to 0.3.1