Add `max_depth` to graph task (#673)
Provide `max_depth` argument in the `@task.graph` decorator, ``@task.graph(max_depth=200)``.  By default, the `max_depth` is 100 nested calls. If the limit is reached, the engine reports a message and raises ``RecursionError``. Deep recursion is generally discouraged, and it is preferred to batch the iteration inside a single task where feasible.

The reported “call depth” is an *approximation* based on the AiiDA process tree, not the exact same call depth as the recursive call.