Attaching extra info to data nodes
Sometimes you need to attach extra information to the data a task creates, like a physical unit or a descriptive label. You can do this directly in the output annotation by passing a dictionary to ``meta(extras=...)``. This populates the ``.extras`` attribute of the resulting AiiDA node, perfect for any custom, JSON-serializable metadata. Let's see an example where we calculate the potential energy of a crystal structure and attach its unit to the output node.