Attaching extra info to data nodes (#717)
In the workgraph's annotation system, one can attach `extras` information for the socket. Originally, this was only used internally to store extra metadata. However, in principle, one can pass any JSON-serializable data into it. This is helpful if one wants to add additional (e.g., semantic) information to the data a task creates, such as a physical unit or a descriptive label. 

This information is stored in the database with the inputs and outputs `spec`, but not attached to the corresponding AiiDA data node. This PR populates the ``.extras`` attribute of the resulting AiiDA node with the `extras` of metadata the user provided. Thus, one can query the data using this extra info.